requirements.txt
# requirements.txt
pandas==2.1.1
requests==2.31.0
psycopg2-binary==2.9.9
apscheduler==3.10.4
python-dotenv==1.0.0
streamlit==1.27.2
plotly==5.17.0
fastapi==0.103.1
uvicorn==0.23.2
